Consolidate tile layers into a single module, `ee_tile_layers`.
